+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 * [#896] Add additional padding to blockquote style so text doesn't blend into decorations
 * [#906] There's always an extra space between special attack name and effect
 * [#918] Actor and Item type names missing from i18n data
+* [#919] Class item template has incorrect default spellcasting class value
 
 ## Chores
 * [#911] Replaced compendium art mapping with default Foundry method.